Refactor Sigmoid
